Remove containers managed by dockermanager and "Created"
This state can only occur when a container was created but could not get
started. This results in the weird behavior that the container is not
running but cannot get started by dockermanager anymore. To prevent that
locking we remove the container and try again.

One case where this can happen is for example that dockerlogstream was
not started yet and the container refuses to start without having the
log driver in place.